Feature: "Move GeneralSubjectList to hostFragment"
- class, sharedViewModel for data sharing purpose in calendar Fragemnt

related to : #147
