+++ app/src/main/java/com/dnd/sixth/lmsservice/presentation/main/classmanage/subject/SubjectFragment.kt
@@ -26,6 +26,7 @@ import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
+import org.koin.androidx.viewmodel.ext.android.sharedViewModel
 import org.koin.androidx.viewmodel.ext.android.viewModel
 import timber.log.Timber
 
@@ -33,7 +34,9 @@ class SubjectFragment : BaseFragment<FragmentClassBinding, SubjectViewModel>(),
     View.OnClickListener, OnRecyclerItemClickListener {
     override val layoutResId: Int
         get() = R.layout.fragment_class
+
     override val viewModel: SubjectViewModel by viewModel()
+    private val hostViewModel: ClassManageViewModel by sharedViewModel()
 
     private var classAdapter: ClassAdapter? = null
     private var viewTreeObserver: ViewTreeObserver? = null
@@ -53,7 +56,6 @@ class SubjectFragment : BaseFragment<FragmentClassBinding, SubjectViewModel>(),
     override fun initActivity() {
         setBindingData() // 필요한 데이터 바인딩
         initView() // 뷰 초기화
-
     }
 
     private fun initView() {
@@ -62,7 +64,7 @@ class SubjectFragment : BaseFragment<FragmentClassBinding, SubjectViewModel>(),
             classAddBtn.setOnClickListener(this@SubjectFragment)
 
             classAdapter = ClassAdapter(
-                viewModel?.generalSubjectDataList?.value!!,
+                hostViewModel?.generalSubjectDataList?.value!!,
                 this@SubjectFragment
             ) // 수업 리사이클러뷰 어댑터
             with(classRecyclerView) {
@@ -70,15 +72,14 @@ class SubjectFragment : BaseFragment<FragmentClassBinding, SubjectViewModel>(),
                 layoutManager = LinearLayoutManager(requireContext())
             }
 
-            // 서버로부터 수업 리스틀를 가져와 업데이트한다.
-            //viewModel?.updateGeneralSubjectList()
+
 
             // 수업 리스트가 변경됨에 따라 화면 크기 조절을 하기 위한 Observer
-            viewModel?.generalSubjectDataList?.observe(this@SubjectFragment) {
+            hostViewModel?.generalSubjectDataList?.observe(this@SubjectFragment) {
                 // ClassManageFragment(ParentFragment)에 수업 개수 전달.
                 ClassManageViewModel.classCount.value = it.size
 
-                if (viewModel?.hasClass() == true) { // 수업이 있다면
+                if (hostViewModel?.hasClass() == true) { // 수업이 있다면
                     //setClassHomeScrollViewHeight() // 수업 RecyclerView Item 사이즈에 맞게 HomeFragment의 Scroll 높이 재설정
                     noClassContainer.visibility = View.GONE // '수업이 없어요' 화면 가리기
                     classRecyclerView.visibility = View.VISIBLE // '수업 목록' 리사이클러뷰 보여주기
@@ -106,16 +107,13 @@ class SubjectFragment : BaseFragment<FragmentClassBinding, SubjectViewModel>(),
                     }
                 }
 
-
-
-
         }
-
 
     }
 
     private fun setBindingData() {
-        binding.viewModel = viewModel // ViewModel 바인딩
+        binding.viewModel = viewModel
+        binding.hostViewModel = hostViewModel // ViewModel 바인딩
     }
 
 
@@ -214,7 +212,7 @@ class SubjectFragment : BaseFragment<FragmentClassBinding, SubjectViewModel>(),
                                 Intent(
                                     requireContext(),
                                     SubjectEditActivity::class.java
-                                ).putExtra("classModel", viewModel.getClassModel(position))
+                                ).putExtra("classModel", hostViewModel.getClassModel(position))
                             )
                             dialog.dismiss() // 하단 Dialog 종료
                         }
@@ -241,7 +239,7 @@ class SubjectFragment : BaseFragment<FragmentClassBinding, SubjectViewModel>(),
             ) { _, _ ->
                 // 수업 삭제 로직 수행
                 CoroutineScope(Dispatchers.IO).launch {
-                    val isSuccess = viewModel.deleteSubject(position) // 수업 삭제
+                    val isSuccess = hostViewModel.deleteSubject(position) // 수업 삭제
                     launch(Dispatchers.Main) {
                         if(isSuccess) {
                             showSnackBar(getString(R.string.success_delete_subject))
